Move drawer expansion logic out of app state
